+++ CHANGELOG.rst
@@ -7,9 +7,11 @@ Change log
 
 .. _Next version: https://github.com/matthiask/feincms3-forms/compare/0.1...main
 
-- Added a special case to ``ConfiguredForm.get_formfields_union``: Passing
-  ``"TYPE"`` in the ``values`` argument returns the lowercased class name of
-  form fields or the type value for ``SimpleField`` proxies.
+- Changed ``ConfiguredForm.get_formfields_union`` to allow fetching arbitrary
+  attributes from plugins.
+- Added a ``TYPE`` classproperty to ``FormFieldBase`` which returns the
+  lowercased class name. This mirrors the ``TYPE`` attribute already existing
+  on ``SimpleFieldBase`` proxies.
 
 
 `0.1`_ (2022-11-10)